requirements.txt
# Google ADK and AI
google-adk>=0.1.0
google-genai>=0.1.0
vertexai>=1.50.0
google-cloud-aiplatform>=1.50.0

# Google Cloud
google-auth>=2.0.0
google-cloud-bigquery>=3.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
typing-extensions>=4.0.0
